[dtraj] Created dtraj package, moved matrix functions io -> util, removed io package.
