+++ msmtools/dtraj/discrete_trajectory.py
