--- msmtools/io/tests/testfiles/matrix_float.dat
@@ -1,3 +1,0 @@
-0.000000000000000000e+00 1.000000000000000000e+00 2.000000000000000000e+00
-3.000000000000000000e+00 4.000000000000000000e+00 5.000000000000000000e+00
-6.000000000000000000e+00 7.000000000000000000e+00 8.000000000000000000e+00
