--- msmtools/io/tests/testfiles/matrix_int.dat
@@ -1,3 +1,0 @@
-0 1 2
-3 4 5
-6 7 8
