--- msmtools/io/tests/testfiles/spmatrix_complex.coo.dat
@@ -1,27 +1,0 @@
- (0.000000000000000000e+00+0.000000000000000000e+00j)  (1.000000000000000000e+00+0.000000000000000000e+00j)  (1.000000000000000000e+01+1.000000000000000000e+01j)
- (1.000000000000000000e+00+0.000000000000000000e+00j)  (0.000000000000000000e+00+0.000000000000000000e+00j)  (2.000000000000000000e+01+2.000000000000000000e+01j)
- (1.000000000000000000e+00+0.000000000000000000e+00j)  (1.000000000000000000e+00+0.000000000000000000e+00j)  (1.000000000000000000e+00+1.000000000000000000e+00j)
- (1.000000000000000000e+00+0.000000000000000000e+00j)  (2.000000000000000000e+00+0.000000000000000000e+00j)  (1.100000000000000000e+01+1.100000000000000000e+01j)
- (2.000000000000000000e+00+0.000000000000000000e+00j)  (1.000000000000000000e+00+0.000000000000000000e+00j)  (2.100000000000000000e+01+2.100000000000000000e+01j)
- (2.000000000000000000e+00+0.000000000000000000e+00j)  (2.000000000000000000e+00+0.000000000000000000e+00j)  (2.000000000000000000e+00+2.000000000000000000e+00j)
- (2.000000000000000000e+00+0.000000000000000000e+00j)  (3.000000000000000000e+00+0.000000000000000000e+00j)  (1.200000000000000000e+01+1.200000000000000000e+01j)
- (3.000000000000000000e+00+0.000000000000000000e+00j)  (2.000000000000000000e+00+0.000000000000000000e+00j)  (2.200000000000000000e+01+2.200000000000000000e+01j)
- (3.000000000000000000e+00+0.000000000000000000e+00j)  (3.000000000000000000e+00+0.000000000000000000e+00j)  (3.000000000000000000e+00+3.000000000000000000e+00j)
- (3.000000000000000000e+00+0.000000000000000000e+00j)  (4.000000000000000000e+00+0.000000000000000000e+00j)  (1.300000000000000000e+01+1.300000000000000000e+01j)
- (4.000000000000000000e+00+0.000000000000000000e+00j)  (3.000000000000000000e+00+0.000000000000000000e+00j)  (2.300000000000000000e+01+2.300000000000000000e+01j)
- (4.000000000000000000e+00+0.000000000000000000e+00j)  (4.000000000000000000e+00+0.000000000000000000e+00j)  (4.000000000000000000e+00+4.000000000000000000e+00j)
- (4.000000000000000000e+00+0.000000000000000000e+00j)  (5.000000000000000000e+00+0.000000000000000000e+00j)  (1.400000000000000000e+01+1.400000000000000000e+01j)
- (5.000000000000000000e+00+0.000000000000000000e+00j)  (4.000000000000000000e+00+0.000000000000000000e+00j)  (2.400000000000000000e+01+2.400000000000000000e+01j)
- (5.000000000000000000e+00+0.000000000000000000e+00j)  (5.000000000000000000e+00+0.000000000000000000e+00j)  (5.000000000000000000e+00+5.000000000000000000e+00j)
- (5.000000000000000000e+00+0.000000000000000000e+00j)  (6.000000000000000000e+00+0.000000000000000000e+00j)  (1.500000000000000000e+01+1.500000000000000000e+01j)
- (6.000000000000000000e+00+0.000000000000000000e+00j)  (5.000000000000000000e+00+0.000000000000000000e+00j)  (2.500000000000000000e+01+2.500000000000000000e+01j)
- (6.000000000000000000e+00+0.000000000000000000e+00j)  (6.000000000000000000e+00+0.000000000000000000e+00j)  (6.000000000000000000e+00+6.000000000000000000e+00j)
- (6.000000000000000000e+00+0.000000000000000000e+00j)  (7.000000000000000000e+00+0.000000000000000000e+00j)  (1.600000000000000000e+01+1.600000000000000000e+01j)
- (7.000000000000000000e+00+0.000000000000000000e+00j)  (6.000000000000000000e+00+0.000000000000000000e+00j)  (2.600000000000000000e+01+2.600000000000000000e+01j)
- (7.000000000000000000e+00+0.000000000000000000e+00j)  (7.000000000000000000e+00+0.000000000000000000e+00j)  (7.000000000000000000e+00+7.000000000000000000e+00j)
- (7.000000000000000000e+00+0.000000000000000000e+00j)  (8.000000000000000000e+00+0.000000000000000000e+00j)  (1.700000000000000000e+01+1.700000000000000000e+01j)
- (8.000000000000000000e+00+0.000000000000000000e+00j)  (7.000000000000000000e+00+0.000000000000000000e+00j)  (2.700000000000000000e+01+2.700000000000000000e+01j)
- (8.000000000000000000e+00+0.000000000000000000e+00j)  (8.000000000000000000e+00+0.000000000000000000e+00j)  (8.000000000000000000e+00+8.000000000000000000e+00j)
- (8.000000000000000000e+00+0.000000000000000000e+00j)  (9.000000000000000000e+00+0.000000000000000000e+00j)  (1.800000000000000000e+01+1.800000000000000000e+01j)
- (9.000000000000000000e+00+0.000000000000000000e+00j)  (8.000000000000000000e+00+0.000000000000000000e+00j)  (2.800000000000000000e+01+2.800000000000000000e+01j)
- (9.000000000000000000e+00+0.000000000000000000e+00j)  (9.000000000000000000e+00+0.000000000000000000e+00j)  (9.000000000000000000e+00+9.000000000000000000e+00j)
