--- msmtools/io/tests/testfiles/spmatrix_int.coo.dat
@@ -1,27 +1,0 @@
-1 1 1
-2 2 2
-3 3 3
-4 4 4
-5 5 5
-6 6 6
-7 7 7
-8 8 8
-9 9 9
-0 1 10
-1 2 11
-2 3 12
-3 4 13
-4 5 14
-5 6 15
-6 7 16
-7 8 17
-8 9 18
-1 0 20
-2 1 21
-3 2 22
-4 3 23
-5 4 24
-6 5 25
-7 6 26
-8 7 27
-9 8 28
