+++ msmtools/util/matrix/__init__.py
