+++ msmtools/util/matrix/matrix.py
