+++ msmtools/util/matrix/matrix_test.py
