+++ msmtools/util/matrix/testfiles/matrix_float.npy
