+++ msmtools/util/matrix/testfiles/spmatrix_float.coo.npy
