+++ msmtools/util/matrix/testfiles/spmatrix_int.coo.dat
